Fix no_auth user creation
Currently, there was an assumption in the tests that allowed servers to
never create the no_auth user based on its token value. This would allow
the services to boot, but have no legitimate user information to use.

This fixes it and removes a hack in the test that obscured the bug. yay.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>