Move to sole owner architecture for managing repos
While github has the notion of multiple owners, we really only use the
notion of an owner for who to run the test as.

This eliminates the multi-owner assumption taken from github and moves
it to single owner. This simplifies the schema and makes the code a lot
less likely to break.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>