Gut and replace crypto strategy with an ideally better one
This replaces the securecookie tooling (which was expiring itself, not
exactly the desired effect of a long-term token store) with a custom
tooling over a CFB+AES cipher provided by golang's standard crypto
packages. It's probably broken in some fundamental way, but we'll figure
that out around release time I guess.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>