types/crypto.go: convert to use GCM from CFB.
Our GCM usage requires no additional data segment, making it an AE
encrypter not an AEAD encrypter (as the documentation would imply). This
is perfectly normal behavior within this system. I checked my
assumptions about carrying the nonce/iv with with the ciphertext; this
seems to be a common strategy so I'm on solid ground there.

This should be a little better than using the CFB cipher suite.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>