In the event a submitted branch is a fork AND added to the db, process it directly.
This change is a little weird. Basically, if you take a parent, fork it,
and then add the *fork* to tinyCI, it would break submitting because it
was trying to resolve the parent, which may not belong to the user.

It was behaving properly, but in the event the fork was submitted
manually, it should behave like an added repository and be tested
directly.

There's probably a few bugs that I haven't caught yet on this feature. I
hope to suss them out soon.

I also added liberal logging to the queue submission in a vain attempt
to figure out why my mocks weren't working, but they seem to add a
little useful tracing data so I'm keeping them for now until real
tracing arrives for real.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>