Support submission identifiers
This feature logs and records submission information in the database, to
be later used by the UI in a future patch. Several refactors and
additional database work was done to accomplish this. More needs to be
done.

The notion of a model submission was created, associations between the
user, the parent, and fork as well as base/head refs were created in the
submission. This allows us to go back and refactor this information out
of tasks to make the database a little more resembling third-normal
form.

Additionally, the notion of a submission in general in the queuesvc
needs to be revisited to make the queuesvc logic simpler. It is not a
replaceable service right now with its overly complicated logic.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>