Reorganize services' import paths
This reorganizes the import paths of the GRPC services so that they can
be easier to track test execution with code coverage tools.

Most of the testservers dir was gutted; but uisvc still remains there
because of the swagger template issues surrounding that. Will visit in a
patch later today.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>