Github readonly client mode
This allows certain workflows with integration tests we want to do, but
ultimately just disallows API requests for certain actions based on a
setting called `readonly_client`, which may be present in the
configuration. If supplied, things like status updates, hook settings
and other "write to github" operations will be nullified, allowing the
system to continue to work otherwise.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>