Fix context use in uisvc
Relying on the context of the request in transactional scnearios lead to
distributed systems problems. Instead, we'll use two contexts; one with
a timeout that can be leveraged by backgrounding routines in case the
connection is interrupted. It has a timeout of one minute for now. We
can make it tunable later.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>