Fix errors.Error use throughout the stack by removing pointer assignment
This patch probably needs a second pass, but this works for now. In
particular, the errors.Error assertions are unwieldy -- the API will
need to change.

This change removes the pointer designation from *errors.Error to be
errors.Error, relying on the interface value of `error` to be nil for
our purposes.

This is much more idiomatic within the go ecosystem and things play
nicer as a result.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>